MOBILE-677: linux compile fix
<==== Mercurial data ====>
Branch: mobile_17.3.1

+++ common/src/utils/media/hevc_sps.cpp
@@ -34,7 +34,7 @@ bool Sps::decodeFromVideoFrame(const QnConstCompressedVideoDataPtr& videoData)
 
         switch (packetHeader.unitType)
         {
-            case (int) hevc::NalUnitType::spsNut:
+            case hevc::NalUnitType::spsNut:
                 return decode(nalu.first, nalu.second);
             default:
                 break;
